Refactor login for testing.
